fix: harness 时间列 Float 改 Double，修复同秒消息排序不稳定
Unix 时间戳(~1.7e9)在 MySQL 单精度 FLOAT 下丢失亚秒精度，
同秒内多条消息 created_at 相同导致 asc/desc 排序不确定，
真实 tool calling 场景 tool 消息可能排在 assistant tool_calls 之前。

- session.py: 6 个时间列(Float -> Double)
- 新增 docker/mysql/migrate_v2_time_precision.sql 迁移脚本
- 已验证：Mock 端到端 created_at 含亚秒精度，user/assistant 顺序正确

+++ crawagent/harness/session.py
@@ -22,7 +22,7 @@ import time
 import uuid
 from typing import Any, Dict, List, Optional
 
-from sqlalchemy import Boolean, Float, String, Text, UniqueConstraint, select, update
+from sqlalchemy import Boolean, Double, String, Text, UniqueConstraint, select, update
 from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, create_async_engine
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -40,8 +40,8 @@ class SessionModel(Base):
 
     id: Mapped[str] = mapped_column(String(32), primary_key=True)
     name: Mapped[str] = mapped_column(String(255), default="")
-    created_at: Mapped[float] = mapped_column(Float, default=0.0)
-    updated_at: Mapped[float] = mapped_column(Float, default=0.0)
+    created_at: Mapped[float] = mapped_column(Double, default=0.0)
+    updated_at: Mapped[float] = mapped_column(Double, default=0.0)
 
 
 class EntryModel(Base):
@@ -56,7 +56,7 @@ class EntryModel(Base):
     tool_calls: Mapped[str] = mapped_column(Text, default="[]")
     tool_call_id: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
     metadata_json: Mapped[str] = mapped_column(Text, default="{}")
-    created_at: Mapped[float] = mapped_column(Float, default=0.0)
+    created_at: Mapped[float] = mapped_column(Double, default=0.0)
 
 
 class LaneModel(Base):
@@ -81,8 +81,8 @@ class OperationLogModel(Base):
     session_id: Mapped[str] = mapped_column(String(32), index=True)
     lane_name: Mapped[str] = mapped_column(String(64))
     op_type: Mapped[str] = mapped_column(String(32))
-    started_at: Mapped[float] = mapped_column(Float, default=0.0)
-    finished_at: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    started_at: Mapped[float] = mapped_column(Double, default=0.0)
+    finished_at: Mapped[Optional[float]] = mapped_column(Double, nullable=True)
     entries_json: Mapped[str] = mapped_column(Text, default="[]")
     error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
 
@@ -98,7 +98,7 @@ class GlobalFactModel(Base):
     session_id: Mapped[str] = mapped_column(String(32), index=True)
     key: Mapped[str] = mapped_column(String(255))
     value: Mapped[str] = mapped_column(Text)
-    updated_at: Mapped[float] = mapped_column(Float, default=0.0)
+    updated_at: Mapped[float] = mapped_column(Double, default=0.0)
 
 
 # ---------------------------------------------------------------------------
